Add remark command
The addressbook can now accept an additional remark command type.

Sometimes the user may want to add remarks to the addressbook
for more clarity but is unable to do so.

The code now accepts the remark command type and displays
the remark when the user inputs a specific remark.

This is preferred for users to input and view details
with greater clarity.