Fix search and sort user stories
The numbering in the ordered list for user stories follows the Markdown
convention of using 1 for all the list items. Furthermore, there was a
typo in one of the lines.

However, in the case of user stories, explicitly specifying the numbers
makes more sense as the numbering is unlikely to change and is
also directly referenced in the extensions section.

Let's change the numbering to specify the step number directly, as well
as fix the typo for clarity.