Logic: Teach the app to accept 'remark' arguments
Modify RemarkCommand to take in an Index and String for remark, and

Modify RemarkCommandTest to test the equals method

Add RemarkCommandParser that will know how to parse two arguments, one
index and one with prefix 'r/'

Add RemarkCommandParserTest that will test different boundary values
for RemarkCommandParser.

Replace the error message with the actual logic to modify the remarks
of a person.