Create patient class
Patient is a subtype of a person and can hence inherit from Person class
and its behaviours.

Creating a patient class along with two additional classes, Condition
and BloodType, which are stored in the patient class, allows us to
abstract different types of people in the system into classes.

Let's make the patient class extend the person class and have it
contain Condition and BloodType so that more information pertaining
to a patient can be stored.

We will add Doctor as a field in the patient class once the Doctor class
is fully implemented.