Modify Delete Class to delete person based on nric
Modified DeleteCommand and DeleteCommandParser classes to delete person
with specified nric, instead of previously which was deleting based on
index. Modified test cases to check for correct deletion based on nric
as well.

Made changes to several other test cases to accomodate for new
isSamePerson method which checks for same nric instead of same name.