chore(gradle): Downgrade JavaFX V21 → 17
This change is in an attempt to improve Java V11 compatibility.
Newer versions of Java V11 seems to be able to handle JavaFX 21,
as is the case for CI,
but some team members are running into JavaFX build issues
on their own computers with Java V11 running.

It is worth noting that the JavaFX docs suggest that both V17 and V21
require Java V17,
yet the project we were given was already upgraded to JavaFX V17
while requiring Java V11 compatibility.