Rectify Edit and Delete implementation
EditCommand was erroneously edited to take personType as a parameter
in the previous commits. Its syntax and implementation remain an issue,
but the errors have been undone at surface level for now.

DeleteCommand was being parsed with a personType, which is unused in
implementation. To resolve this, ParserBasic and ParserComplex classes
have been written to better suit the needs as well as aid in testing.
- Basic commands do not require specifying personType
- Complex commands require a personType tag
It is very likely that we will add more basic commands in the future as
well, for example changing user settings, exporting data etc.

Also fix some bugs with the parsing of different command Patterns
- Previously Basic commands could also be parsed as Complex
- This has been fixed by altering the Complex command regex to be more
restrictive.