Update patient class previously created
Added additional fields such as emergency contact which was missing
previously but essential to the patient class. Doctors field will be
added later on after Doctor class is created.

Additionally, Patient specific details were not shown in the GUI
previously. Many patient specific files were created as I was not
able to create a person observable list that contains patients.
This also means that doctors specific files need to be created
similarly.

In the future we also need to find a way to show both doctors and
patients in the GUI. Currently patient list replaces the person list.
Eventually doctors list wil need to be added.