Fix data representation error and add personProfilePlaceholder StackPane
Currently the PersonProfile FXML has issues displaying tags which causes the error.
Fix the tag display to prevent the error.
Using the already existing setVisible() method of StackPane can toggle the visibility of the fosterer list and the profile view.
Add a personProfilePlaceholder StackPane below the personListPlaceholder StackPane and allow the visibility to toggle between them.