Add an Appointment field to Person and implement ScheduleCommand.
Previously, person contact did not have have an appointment field. There was no command to schedule appointment.

Add the Appointment class for appointments made and NullAppointment subclass for empty appointments. Added command to add appointment for a person.

This is done so as to allow user to schedule appointment for his clients