Add ability to edit persons to events
Update edit_meeting command: If editing the event with a valid name, it will add the person name to the existing associated persons. E.g: if an event (1) has Bob already assigned to it, edit_meeting 1 n/John would add John to the list instead of replacing it. This makes the adding of associated persons easier.
Update corresponding success and failure messages.