Add confirmation page for clear command
The clear command is not protected by any validation. This
makes new users vulnerable to deleting their contact book while learning
how to use it.

Adding a confirmation page allows users to stop and think about the
ramifications of their actions and reduces the likelihood of mistakes
that can worsen the user experience.

Let's add a confirmation page.