Fix edit command bug
Currently, edit fails even though an attribute of Person is present and
user likes to change to another value.

The equals method of Person is not updated accordingly; it does not
take into account other alternative contact information added.

Equals method in Person is updated and test cases are added to
maintain code coverage.