Implement PayRate
Person class does not have a field to keep track of earnings
for user (private tutor).

Add PayRate class and modify Person and corresponding commands
and parser classes to work with new field.

PayRate class has a int value so that we can implement
command that calculates total earnings in the future.