Handle duplication that might occur in adding or editing
Add and edit commands have been changed such that a new entry or an edit to an entry cannot use a policy number used by another code. This check should not occur for code with no policy.

The issue of 2 policies with the same name has also been ammended. Instead, addition or edits should work as long as all the fields are not the same. This is the only duplicate case.

Some of the test cases in EditCommandTest are yet to be handled.