Add FindTeam and FindTeamLeader command classes
The current implementation lacks functionality to specifically find a
team and its leader by their respective attributes.

To enhance the search capabilities of the application, it's necessary
to provide commands that allow users to efficiently retrieve teams and
their leaders based on given criteria.

Let's,
* add the `FindTeamCommand` and `FindTeamLeaderCommand` classes
* implement parsers for both commands to handle user input
* extend the `Messages` class to support new error and success messages
  for these commands

The decision to have separate classes for each command ensures modularity
and separation of concerns. It also paves the way for potential future
enhancements specific to each command.