Implement Sort Command
The use of "sort" will be recognized as a command to arrange persons from closest expiring policy expiration date to latest.

Those with no specified policy are at the end of the database.

The persons are not auto sorted and thus everytime you run the sort command must be used again.