Refactor DateTime into Date and Time
The DateTime class currently handles both date and time operations.

To avoid Date repetition in the start and end of an Appointment since
an Appointment will start and end on the same Date, we separate Date
and Time.

Splitting Date and Time would allow Appointment to hold one Date, with
two Time to indicate the startTime and endTime of an Appointment.

This change is made to enhance clarity and reduce redundancy in the
Appointment class.