Add prompt for updating appointment
Currently, when scheduling an appointment with a person with an appointment
scheduled, the schedule command overrides the old appointment without
the users knowledge.

This fix keeps the implementation of this overriding appointment but adds a
prompt that lets users know the current appointment scheduled and asks the
user if he would like to override the appointment

This is done to enable schedule to have dual functionality of adding and
editting as well as a reminder of a potentially forgotten appointment with
a person and also as a quick way to override past appointments with a new
one.