Modify Find Command to accept financial plans and tags
Currently, the find command is rather limited in scope. It is only able
to search for names while a user might want to search by financial plans
or tags.

Let's:
- Expand the find command to search by tags and financial plans
  (substrings included)

We choose to search names by whole words because it is unlikely to
search for names by substrings, while tags and financial plans may have
substrings that make sense to search for.

Testing will be added in a future commit.