Add find by module
Current implementation of find command only
allows find by name.

To allow users to easily filter TAs by
a specific module.

Modifying the FindCommandParser to check for
the mod prefix, and getting the specific
mods to filter the TAs.

It allows us to search through the list of
TAs and find those that are teaching those
mods.