Add CombinedPredicate class and tests for previous commit
The current predicate implementation makes it difficult to determine
if two predicates for a Person are the same. This implementation makes
it slightly easier to compare, such that as long as the keywords within
each component predicate are ordered in a certain way, then two
CombinedPredicates will be the same. It's good enough for the current
purpose anyway.