Add Panel for Job Details
Current view of the list of job application is getting too long as more fields are added.

Implement UI components to support the Job Details Panel indicate details of a job application.

So that only essential details of the application is put at the JobList Panel. While the rest are shown on the JobDetailsPanel.

- Implement logic in MainWindow.java to displayJobDetails when prompted
- Configure FXML component and css styling for the newly added component.