Add new UndoCommand
The command currently does not work still. There is an issue with the
copying of Files which the rest can hopefully figure out.

Here is how the command works:
When new command is added, model.saveToFile is called which will copy
the current file into the previous file.
When undo command is called, model.setToPreviousFile is called which
will copy the previous file into the current file, hence "reverting"
the changes.