Modify message usage for FindCommand
Current FindCommand message usage does not reflect
new changes made to the command.

To let the user know how to use the new features
we added to the FindCommand.

Add more examples of how to use the command.

To let the users have a better idea of how to use
the command.