Add COMMAND_ALIAS for commands
Right now, many command names are long and inconvenient to type. I added a constant COMMAND_ALIAS in these command classes and AddressBookParser would recognise both it and COMMAND_WORD to return the same command.

We may have to change some messages and the UG to reflect and highlight this change.