Update find command
The UG does not have any details about the find command, which
allows users to find teaching assistants with certain search
parameters, such as name, course and free time.

Let's include the format, examples and use cases of this find
command, to give users a better idea of how to use this command.