Change find command implementation
Previously find command was searching by words, where there are no partial searches allowed
Eg. searching "abc" would not find "abcde".

Also, it is not possible to search using a string with spaces, since it uses words to search, where spaces separate words.

This should change since it is not the intuitive idea of finding for students.

The find command is changed to allow partial searching of keywords and the specified string is taken as a entire string.