Edit output message of undo/redo commands for Enrol specific commands
The current behaviour of undo/redo command for Enrol specific command is wrong as
it outputs a cryptic looking message upon undoing/redoing the Enrol command. This
is due to EnrolCommand class using the toString method of the Attendance class to
output the message.

A new method is created for Attendance class, which is called getMemberAndEventAttendance.
This method formats the output message to be more user-friendly and less cryptic. The
EnrolCommand then calls on this method to display the more user-friendly message.

Doing so will make the output message clearer and does not expose the internal structure
of the code to the user.