Tree Command: Display LinkTree in A New Window
Our hierarchy of contact information are not visualizable.

Visualization of hierachy of contact information aids the user to make
decisions.

Visualizing the tree structure provides a general hierarchical view of
all the contact information, which helps users to decide who to
contact.

Let's have a "tree" command that allows users to see the tree
structure in a new window.

Displaying in a new window allows bigger display of the tree structure
so that users can navigate easily.