Integrate CommandHistory into CommandBox
CommandHistory class was created but not used in CCACommander, thereby
making it unnecessary. There was also a lack of functions in
CommandHistory that facilitates the feature of recalling past commands.

Specification is as follows:
Any command that is not the same as the last command stored into the
command history will be stored into commandHistoryList. Upon
pressing the up arrow, the user will be able to get the previous
command and display it in the command box. Similar with the down arrow
and the next command.

There is a special case of deletion. When the user presses the
backspace, it will reset the pointer and only allow the user to
recall the previous commands.

Let's include
* resetPointer() function that resets the pointer as the size of
  commands, this is past the last command.
* isLastCommand() function to return whether the pointer is at the last
  command.
* CaseSwitch statements in CommandBox based on keyPress
  * Up Arrow to get previous command
  * Down Arrow to get next command
  * BackSpace to resetPointer
* AddCommand when entering command into command box