Fix bugs related to UndoCommand and RedoCommand
UndoCommand and RedoCommand did not update the UI correct as intended
as setAddressBook was not utilised. Instead, addressBook was assigned
to the previous addressBook normally which led to the UI not receiving
the feedback and updating accordingly.

RedoCommand did not work as well due to it referencing the same
addressBook instance instead of creating a new one each time, thus redo
simply lead to an empty addressBook being set each time.

Made the fixes necessary to correct the bugs. Also updated test cases to
test for situations where the redo and undo commands are called over the
maximum of 5 times, which should return the error message.