Fix EditCommand Issue, add SaveCommand, and add field focus functionality
EditCommand is currently having issues with setting every field to nil.
Fix the EditCommand issue so that it properly sets the field values.
Add SaveCommand so that user can save the changes made in the profile view page.
Add field focus functionality in the profile view page for better user experience.