Replace assertion with Exception for Appointment
Currently, Appointment uses Strings to store its values, which
allows assertions to easily validate the state of the internal value
through the use of RegEx.

However, Strings are very limiting in storing Appointment, as it
prevents more stringent enforcement of conditions for the values.
In particular, Strings make it hard to enforce the natural ordering
properties of the start and end times, and the validity of the input
particularly for days in a month. This necessitates the use of
DateTime objects, which only checks for the validity of an input
during a parse of DateTime objects - in every situation we choose
to check for input validity, we will always produce a DateTime
object as a side effect if the input is valid, and an Exception if
it is not.

Therefore, we opt to change Appointment parsing to handle the
exceptional cases in either ParserUtil or Appointment, rather than
enforce only valid inputs into Appointment.