Add courses json file to data directory
Course data is currently not being stored in a human
editable file.

We need to store course data, such as lessons, in a
human editable file.

Created a courses.json file in the data directory for
users to modify.

To ensure there is a human editable file for users to
modify course data.