Add AddAppointmentCommand
This commit adds the AddAppointmentCommand to the project. It also
updated the required UI changes to display the appointment in each
patient/doctor card. Users should be able to add appointments to the
address book and see the newly added appointments in the UI.

However, the tests are currently failing. It will be fixed in the next
PRs, either by me or by my teammates. Additionally, the current way of
writing the code may be violating coding principles and may require
changes. I will be looking into it in the next PRs and also discuss
with my teammates.