Fix Edit Musician Bug
There was a bug in the edit musician method which resulted in
the new musician being added to all bands stored in the addressbook
regardless of if they were there beforehand.

This bug has been fixed to only store the edited musician if they
already existed in a particular band prior to the edit.