Add save confirmation before exiting PersonProfile and prevent potential errors due to SaveCommand
The user should be asked to save the edited details before making a change.
Add save confirmation so that users are informed to save the changes before exiting.

Also, when the user adds a new fosterer in the system, SaveCommand should throw an error if the user did not fill in the details.
Add checks so that SaveCommand does not cause malicious errors.