Improve code quality and add custom feedbacks for different add command use cases
Currently mainWindow code quality can be improved with the use of CommandType parameter in CommandResult.
Improve code quality of mainWindow by removing unnecessary null check for EditFieldCommand, which may also potentially cause troubles.

Currently adding from the profile page prompts the same confirmation message "You did not save your changes" even when saving is disabled when not every field is filled in.
Add a message for the case when not every field is filled in and the user wants to exit the profile page.