Fix minor bug on events
There was a problem on the add event and delete event commands, where they don't work when there is no end time. This is because one of the functions used does not allow null values.

To solve this, inside the function was made so that the end time would be equal to the start time when it is null.