Refactor CombinedPredicate to be more flexible
CombinedPredicate only accepts one of
FinancialPlanContainsKeywordsPredicate, NameContainsKeywordsPredicate,
and TagContainsKeywordsPredicate each. This makes the usage of
CombinedPredicate difficult and inflexible. Using a common interface
with varargs will make the usage of CombinedPredicate easier and
more flexible.

Let's:
- Refactor the CombinedPredicate class
- Introduce a PersonContainsKeywordsPredicate interface to abstract the
  other predicate classes

A limitation to this current implementation would be that
CombinedPredicate is a bit of a misnomer, since it is an "or" predicate.
An "and" predicate of a similar nature would cause a naming clash.
However, we don't need such a predicate yet, so we'll stick to this
naming.
Also, there might be concerns about the ability of this
new CombinedPredicate to compare against other objects
of the same type. We don't need to
compare predicates by their actual functionality now,
so implementing a stronger equals() method can be done later.