Fix date string parser method
Before, in the case where the string contains a '-', it doesn't check whether it is in the yyyy-MM-dd format.

This results in weird error messages. For example, when the command is . This would have an error message 'Text '2022-01-02 100 00:00:00' could not be parsed...'. Because, we were previously assuming that a string containing '-' is in the yyyy-MM-dd format, where we want to automatically append 00:00:00.

This is the same for the other case (Only Time).