Fix load bug
Initially, an error in the JSON file caused an empty Addressbook to be loaded.
This can cause the user to be confused as to what can be done, thus we made a sample data be loaded
when a bug occurred.

This fixes the AB3 requirement where a sample addressbook should be loaded when the
save file is corrupted