Bug fix: wrong reading information
The app will have a NullPointerException if we add the last person
to a team and delete a person before him, then close the app and
restart it. Person's Identity code will not be read correctly.

The exception will cause the whole app to crash.

Let's fix this bug by making the Persons' Identity codes be the
same as what is stored in the file when reading information from
the file.

In this way, there will not be a NullPointerException in this
situation.