Fix error message of invalid index of entry
Before, when a number greater than `Integer.MAX_VALUE` is provided
as index of entry for a multi-valued field, the error message was
"Index is not a non-zero unsigned integer". The error message was wrong
because the index could still be a non-zero unsigned integer.

Now, the error message will be consistent with index of contact,
which will be "invalid command" followed by command usage.