Fix bug stopping ManageHR from launch
When an employee has supervisors that are added at a later stage, the app
will fail to relaunch when it is restarted.

This prevents the user from using the app at all.

In order to address this problem, the app first reads data from the JSON file
without enforcing any constraints. Only when all the data is loaded,
will the constraints be enforced.

The issue lies in ensuring that the supervisors need to exist
before any employees can be under the charge of that employee.
Since supervisors could also locate be located at the tail-end of the
data file, our current implementation fails to take this into account.