Bugfix: add and edit using tags that haven't been created
When doing `edit 1 t/newcat test`, it should add the new tag to the
person at index 1 and add the new tag to the tag list. Bug fix is to
to add the create a new tag, add it to list and return it if the
current tag list does not have the specified tag.